refactor: Update DTOs name
- DTO명 엔티티와 구분 쉽도록 변경

+++ src/main/java/com/cherrydev/cherrymarketbe/server/domain/order/dto/responses/OrderCreateResponse.java
@@ -3,6 +3,6 @@ package com.cherrydev.cherrymarketbe.server.domain.order.dto.responses;
 public record OrderCreateResponse(
         String orderId,
         String orderName,
-        Integer amount
+        Long amount
 ) {
 }
